Move various chain type traits to chain-type-components
