+++ crates/chain/chain-components/Cargo.toml
@@ -12,6 +12,7 @@ description  = """
 """
 
 [dependencies]
-cgp                         = { workspace = true }
-hermes-runtime-components   = { workspace = true }
-hermes-encoding-components  = { workspace = true }
+cgp                             = { workspace = true }
+hermes-runtime-components       = { workspace = true }
+hermes-encoding-components      = { workspace = true }
+hermes-chain-type-components    = { workspace = true }
