+++ crates/chain/chain-components/src/traits/types/height.rs
@@ -2,28 +2,9 @@
    Trait definition for [`HasHeightType`].
 */
 
-use core::fmt::Display;
-
 use cgp::prelude::*;
 
-#[derive_component(HeightTypeComponent, ProvideHeightType<Chain>)]
-pub trait HasHeightType: Async {
-    /**
-       The height of the chain, which should behave like natural numbers.
-
-       By default, the height only contains the `Ord` constraint, and does
-       not support operations like addition.
-
-       We can impose additional constraints at the use site of `HasChainTypes`.
-       However doing so may impose limitations on which concrete types
-       the `Height` type can be.
-
-       By keeping the abstract type minimal, we can for example use
-       `u8` or `u128` as the `Height` type during testing, and use the
-       more complex Cosmos height type during production.
-    */
-    type Height: Ord + Display + Async + Clone;
-}
+pub use hermes_chain_type_components::traits::height::*;
 
 #[derive_component(HeightFieldComponent, HeightFieldGetter<Chain>)]
 pub trait HasHeightFields: HasHeightType {
