+++ crates/chain/chain-type-components/src/lib.rs
@@ -1,1 +1,3 @@
 #![no_std]
+
+pub mod traits;
