+++ crates/chain/chain-type-components/src/traits/ibc/connection_id.rs
@@ -1,0 +1,12 @@
+use core::fmt::{Debug, Display};
+
+use cgp::prelude::*;
+
+#[derive_component(ConnectionIdTypeComponent, ProvideConnectionIdType<Chain>)]
+pub trait HasConnectionIdType<Counterparty>: Async {
+    /**
+       The connection ID of the counterparty chain, that is stored on the self
+       chain.
+    */
+    type ConnectionId: Debug + Display + Async;
+}
