+++ crates/chain/chain-type-components/src/traits/ibc/mod.rs
@@ -1,0 +1,5 @@
+pub mod channel_id;
+pub mod client_id;
+pub mod connection_id;
+pub mod port_id;
+pub mod sequence;
