+++ crates/chain/chain-type-components/src/traits/mod.rs
@@ -1,0 +1,4 @@
+pub mod event;
+pub mod height;
+pub mod ibc;
+pub mod message;
